updated as per discussion
