+++ src/reporter/tests/test_NTNENA.py
@@ -1,6 +1,6 @@
 from conftest import QL_URL
 from datetime import datetime
-from reporter.tests.utils import delete_test_data, insert_test_data, insert_test_data_different_types
+from reporter.tests.utils import delete_test_data, insert_test_data
 import pytest
 import requests
 import dateutil.parser
@@ -31,7 +31,7 @@ def query(values=False, params=None, service=None):
     return requests.get(query_url(values), params=params, headers=h)
 
 
-@pytest.fixture(scope='function')
+@pytest.fixture(scope='module')
 def reporter_dataset():
     for service in services:
         insert_test_data(service, [entity_type], n_entities=1, index_size=4,
@@ -44,16 +44,6 @@ def reporter_dataset():
     yield
     for service in services:
         delete_test_data(service, [entity_type, entity_type_1])
-
-@pytest.fixture(scope='function')
-def reporter_dataset_different_types():
-    for service in services:
-        insert_test_data_different_types(service, [entity_type], n_entities=1, index_size=4,
-                         entity_id=entity_id)
-
-    yield
-    for service in services:
-        delete_test_data(service, [entity_type])
 
 
 @pytest.mark.parametrize("service", services)
@@ -1200,333 +1190,3 @@ def test_aggregation_is_per_instance(service, reporter_dataset):
         'attrs': expected_attrs
     }
     assert obtained == expected
-
-
-def test_aggregation_different_types_timescale(reporter_dataset_different_types, service = 't1'):
-    attrs='temperature,intensity,boolean'
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'min'
-    }
-    # timescale do not support min on boolean
-    r = query(params=query_params, service=service)
-    assert r.status_code == 404, r.text
-
-    obtained = r.json()
-    assert isinstance(obtained, dict)
-
-    expected = 'AggrMethod cannot be applied'
-
-    assert obtained['error'] == expected
-
-    # 'aggrMethod': 'max'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'max',
-    }
-
-    # timescale do not support max on boolean
-    r = query(params=query_params, service=service)
-    assert r.status_code == 404, r.text
-
-    obtained = r.json()
-    assert isinstance(obtained, dict)
-
-    expected = 'AggrMethod cannot be applied'
-
-    assert obtained['error'] == expected
-
-    # 'aggrMethod': 'count'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'count',
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 200, r.text
-
-    obtained = r.json()
-    assert isinstance(obtained, dict)
-    expected_attrs = [
-        {
-            'attrName': 'boolean',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [4]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'intensity',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [4]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'temperature',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [4]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-    ]
-
-    expected = {
-        'attrs': expected_attrs
-    }
-
-    obtained = r.json()
-    assert obtained == expected
-
-    # 'aggrMethod': 'avg'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'avg'
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 404, r.text
-    assert r.json() == {
-        "error": "AggrMethod cannot be applied",
-        "description": "AggrMethod cannot be applied on type TEXT and BOOLEAN."
-    }
-
-    # 'aggrMethod': 'sum'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'sum'
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 404, r.text
-    assert r.json() == {
-        "error": "AggrMethod cannot be applied",
-        "description": "AggrMethod cannot be applied on type TEXT and BOOLEAN."
-    }
-
-
-def test_aggregation_different_types_crate( reporter_dataset_different_types, service = 't2'):
-    attrs='temperature,intensity,boolean'
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'min',
-        'type': 'Room'
-    }
-    # crate supports min on boolean
-    r = query(params=query_params, service=service)
-    assert r.status_code == 200, r.text
-
-    obtained = r.json()
-    assert isinstance(obtained, dict)
-
-    expected_attrs = [
-        {
-            'attrName': 'boolean',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [False]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'intensity',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': ['str1']
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'temperature',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [0.0]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-    ]
-
-    expected = {
-        'attrs': expected_attrs
-    }
-
-    assert obtained == expected
-
-    # 'aggrMethod': 'max'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'max',
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 200, r.text
-
-    obtained = r.json()
-    assert isinstance(obtained, dict)
-    expected_attrs = [
-        {
-            'attrName': 'boolean',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [True]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'intensity',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': ['str1']
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'temperature',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [3.0]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-    ]
-
-    expected = {
-        'attrs': expected_attrs
-    }
-    obtained = r.json()
-    assert obtained == expected
-
-    # 'aggrMethod': 'count'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'count',
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 200, r.text
-
-    obtained = r.json()
-    assert isinstance(obtained, dict)
-    expected_attrs = [
-        {
-            'attrName': 'boolean',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [4]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'intensity',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [4]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-        {
-            'attrName': 'temperature',
-            'types':
-            [{
-                'entities':
-                [{
-                    'entityId': 'Room1',
-                    'index': ['',''],
-                    'values': [4]
-                }],
-                'entityType': 'Room'
-            }]
-        },
-    ]
-
-    expected = {
-        'attrs': expected_attrs
-    }
-
-    obtained = r.json()
-    assert obtained == expected
-
-    # 'aggrMethod': 'avg'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'avg'
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 404, r.text
-    assert r.json() == {
-        "error": "AggrMethod cannot be applied",
-        "description": "AggrMethod cannot be applied on type TEXT and BOOLEAN."
-    }
-
-    # 'aggrMethod': 'sum'
-
-    query_params = {
-        'attrs': attrs,
-        'aggrMethod': 'sum'
-    }
-    r = query(params=query_params, service=service)
-    assert r.status_code == 404, r.text
-    assert r.json() == {
-        "error": "AggrMethod cannot be applied",
-        "description": "AggrMethod cannot be applied on type TEXT and BOOLEAN."
-    }
